check navbar for every navigation
